move and update files
