+++ inc/Dependencies/League/Container/Inflector/Inflector.php
