+++ inc/Dependencies/League/Container/Inflector/InflectorAggregate.php
