+++ inc/Dependencies/League/Container/Inflector/InflectorAggregateInterface.php
