+++ inc/Dependencies/League/Container/Inflector/InflectorInterface.php
